front end prototype
Included is the header and footer that will be present on each page. right now they are in their own file as google web services let me implement them without combination in each file, but we can go to a more traditional implementation if we want. Also the registration and sign in page are done for the front end. let me know if anything looks funky or weird as I haven't used github for this before. right now the index.html page is currently the sign in page.