Gun1.2
Everything here should be pulled, but I am unsure of changes made in Django syntax in the HTML files, so I will leave that to the pros. There are a lot of changes in most of the HTML files. Created an Account management page that currently can be accessed by selecting the "MyCoin1" in the ICO page as a temporary place holder until we've worked out the log in functionality. I  created a coin template page that can be navigated to by selecting any coin in the market page or ICO page. implemented a rough version of the graphs for the site. This graph is subject to change, but should work well enough for the first iteration. We will need to pull the new style.css and make sure to add account.html to the template folder. I am an idiot when it comes to github or I would do these changes myself. Thanks Men.